+++ tests/sources/quasar/sfpu_fill_quasar_test.cpp
@@ -78,11 +78,11 @@ const bool is_int_fpu_en = false;
 
 #include "cfg_defines.h"
 #include "cmath_common.h"
+#include "experimental/ckernel_sfpu_fill.h"
 #include "llk_math_common.h"
 #include "llk_math_eltwise_unary_datacopy.h"
 #include "llk_math_eltwise_unary_sfpu_common.h"
 #include "params.h"
-#include "sfpu/ckernel_sfpu_fill.h"
 
 using namespace ckernel;
 using namespace ckernel::math;
